fix small changes in diagram
