+++ .idea/dictionaries/darta.xml
@@ -1,14 +1,18 @@
 <component name="ProjectDictionaryState">
   <dictionary name="darta">
     <words>
+      <w>bens</w>
+      <w>darta</w>
       <w>dirc</w>
       <w>elwl</w>
       <w>geoew</w>
       <w>geons</w>
       <w>hdway</w>
       <w>imgs</w>
+      <w>jucson</w>
       <w>junc</w>
       <w>nlsl</w>
+      <w>outjunc</w>
       <w>phsr</w>
       <w>phsrlst</w>
       <w>pptx</w>
