+++ mydata.json
@@ -1,0 +1,1 @@
+{"tux": {"health": 23, "level": 4}, "beastie": {"health": 13, "level": 6}, "konqi": {"health": 18, "level": 7}}
